+++ src/main/java/tensor/TensorDataTypeUtil.java
@@ -9,7 +9,7 @@ final class TensorDataTypeUtil {
         }
         if (left == DataType.FLOAT64 || right == DataType.FLOAT64) return DataType.FLOAT64;
         if (left == DataType.FLOAT32 || right == DataType.FLOAT32) return DataType.FLOAT32;
-        return DataType.FLOAT16;
+        return DataType.BFLOAT16;
     }
 
     static DataType binary(Tensor first, Tensor second) {
